requirements.txt
Django==1.8.3
django-clear-cache==0.3
django-filter==0.10.0
django-suit==0.2.13
djangorestframework==3.1.3
djangorestframework-camel-case==0.2.0
Markdown==2.6.2
MarkupSafe==0.23
Pillow==2.9.0
sendgrid==1.4.0
sendgrid-django==1.3.0
smtpapi==0.2.0
wheel==0.24.0
